requirements.txt
faiss-cpu==1.11.0.post1
fitz==0.0.1.dev2
google-generativeai==0.8.5
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
pandas==2.3.1
sentence-transformers==5.1.0
streamlit==1.48.0
tqdm==4.67.1
